+++ include/server/asio/udp_client.h
@@ -185,8 +185,8 @@ private:
     asio::ip::udp::socket _socket;
     std::atomic<bool> _connected;
     // Client statistic
-    size_t _bytes_sent;
-    size_t _bytes_received;
+    uint64_t _bytes_sent;
+    uint64_t _bytes_received;
     // Receive endpoint
     asio::ip::udp::endpoint _recive_endpoint;
     // Receive & send buffers
